Fixes address decomposition and shift instructions
Corrects address decomposition logic to ensure proper bit extraction and recomposition, which is critical for memory addressing in FHEVM.

Updates shift instructions to prevent out-of-bounds shifts, ensuring correct data manipulation.

Add test for loading instruction

+++ fhevm/tests/decompose.rs
@@ -15,7 +15,7 @@ fn decompose_u32() {
 
     let value: u32 = 0xf0f0f0ff;
 
-    let result: Vec<i64> = decomposer.decompose(&module_pbs, &precomp, value);
+    let result: Vec<u8> = decomposer.decompose(&module_pbs, &precomp, value);
 
     let mut have: u32 = 0;
